Rename model_object
